third_party/WebKit/LayoutTests/css-parser/unitless-zero-angle.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/property-parsing-test.js"></script>
<script>
assert_valid_value("image-orientation", "0deg");
assert_valid_value("rotate", "0deg");
assert_valid_value("offset-path", "ray(0deg farthest-corner)");

assert_invalid_value("image-orientation", "0");
assert_invalid_value("rotate", "0");
assert_invalid_value("offset-path", "ray(0 farthest-corner)");
</script>
